requirements.txt
flask
flask-login
werkzeug
psycopg2-binary
requests
pandas
python-dotenv
